Added support for perpetual simulations
Input parameter solday in rrtm_radiation_nml can set the day of the year to a constant value.
If solday >0, the day of the year will always be set to solday.
In order to set perpetual NH/SH winter solsitce, set solday = 360/180
For equinoxes, set to 90/270.
It is also possible to set to fractual day, eg 0.5 or 180.25, to set a specific time of the day.