Found a potentially easier way to include perpetual simulations
The only place where the perpetual or seasonal cycle calculations differ is in run_rrtmg.
There is one check for solday>0, and if so, it sets Time_loc to the solday and seconds therefrom.
Time_loc is used for compute_zenith, and the interpolator for ozone.
For everything else, Time is used (ie diagnostics, time step check, etc)