Make albedo a function of precipitation
Whereever precipitation is positive, set albedo to precip_albedo. This is only done if do_precip_albedo = .true.
Precipitation is passed from moist_processes to run_rrtmg via the module rrtm_vars.
As radiation is not called at every time step, an average 'time of precipitation' is computed in
moist_processes. In run_rrtm, this average is used to weight the albedo from 'albedo' towards 'precip_albedo'.
Thus, albedo_loc is only equal to precip_albedo if it has been raining continuously between two calls to
RRTM (and who wants that, anyway?).
Conversely, it is only equal to the albedo from simple_surface if it has never been raining anywhere in the
corresponding grid box (which doesn't happen often either).