Bugfix for small coszen
Values of cosine of zenith angle around 1.e-3 seem to cause problems in RRTM.
Therefore, values below 1.e-2 are now set to zero. The is equivalent
to neglecting the effect of sunlight where there is less than 15min
of sunlight per day. Probably not a problem, given that the time step is
of the same order of magnitude.