Added Q-flux
In addition to the already implemented oflx (ENSO) and oflxmerid (zonally symmetric) ocean fluxes,
added Tim Merlis' (2013) qflux, which is based on a Gaussian around the equator.
The important difference to oflxmerid is, that it is described analytically and can therefore be
modified with input parameters, whereas oflxmerid is fixex coded in with a lookup table.