SST as initial condition
do_sc_sst is a new flag, which sets whether or not the read sst
(if do_read_sst=.true.) should be used as initial condition only,
or as a time dependent climatological input.
Also, if surface_choice/=1, the sst will not be read as a function
of time, regardless do_sc_sst. This, surface_choice=2 will always
result in fixed SST in time.