Choice of day of the year for equinoxes.
equinox_day \in [0,1] gives the fraction of the year that marks the spring equinox.
By default, equinox_day = 0.25 defines the spring equinox at one fourth of the
duration of the year.
For starting the simulation with equinox conditions, set equinox_day = 0.

However, be aware that equinox_day /= 0.25 means that the seasons will not correspond
to the standard output seasons, ie for equinox_day = 0. month 1 will be April.