Zonal mean water vapor / Critical latitude for precip_albedo
With do_zm_tracers=.true. , only the zonal mean of water vapor is fed to RRTM.
It's only water vapor, as CO2 is one scalar, and O3 can be read from a file,
which in turn can be set to be zonally symmetric.

precip_lat defines poleward of which latitude the precipiation albedo is applied
if do_precip_albedo = .true.