Allow for thirty day calendar interpolation
So far, input files had to be either climatological, or
have Gregorian or Julian calendars.
Now, in preparation to allowing more input files for MiMA,
the interpolator also accepts THIRTY_DAY_MONTHS, provided
both the file and the code are on the same calendar.
In particular, this will allow to feed the code its own
previous output as input.