Clarification of use_dyofyr
The use_dyofyr flag chooses whether or not the day
of the year should be used to compute the Earth-Sun
distance. This is native to RRTM, but assumes a 365
day year. Thus, there is now a check if the year
really has 365 days before accepting the choice of
use_dyofyr=.true.