Added radiation options and license info
Gray radiation can be run instead of RRTM, with the appropriate flags. There is also a cautionary note regarding the possibility of running the full AM2 radiation.
Finally, AM2 and MOM, of which part are included in MiMA, are also distributed under a GNU GPL license. This info has been added.