Re-arrangement of solday check
Within rrtm_radiation_init, the check for perpetual simulations was
done before the namelist value of solday was read. This had no
influence on the simulation output, but the more saying that a
perpetual simulation is run could have been triggered (or not)
independently of the actual value of soldiery. Fixed now.