Introducing np_cap_factor
In simple_surface, np_cap_factor changes the heat_capacity in the
northern hemisphere. With this, it is possible to adjust the
extratropical heat capacity in the northern hemisphere, to be different
to the southern hemispheric heat capacity. This is useful to induce a
larger (or smaller) seasonal cycle in baroclinicity in the northern
hemisphere, without the abrupt changes when using land_capacity.