Scaling ozone input
Instead of having to create a new input file,
there is now a possibility of scaling the values
within the ozone input file by a constant factor.
This can be useful e.g. when performing a parameter
sweep of ozone concentration, or to slightly adjust
the ozone concentrations coming from the input file.