End of day
There seems to be a memory leak/bug somewhere.
The variable 'tt' within astro.f90 has random errors
(they change from run to run) if topography_option = 'input'
with the topography file orog.T42.nc.
This does not happen with the option 'gaussian', and it
does not happen with topo_poscos_4000.nc.
However, there is no reason why the topography file should
change any variable in astro.f90, so it must be a memory thing.
In addition, no problems come up if running with 4 CPUS, but they
do appear with more CPUs, and become more frequent (spatially)
as well.
To be continued.