Land sea mask as input file
It is now possible to read in a land-sea
mask from an input file, and apply
land_capacity whereever the field within
the file is > 0.
Typically, the land sea mask would thus be
1 over land and 0 over sea.
As an example, added lmask.T42.nc, which is
1 where the surface geopotential of the GFDL-CM3
land sea mask (interpolated onto the T42 grid)
is larger than 50 m2/s2.