Updated recommended values
The vertical coordinate going to 11 scale heights was a safe choice, but probably not the best choice.
11 scale heights is approx. 0.001hPa, and the winds are are very strong at those heights. Thus, a very short time step is needed.
The new value, 9 scale heights, sets the top at around 0.01hPa, which is still high enough to be a high top model, but it allows for a longer time step. Thus, the time step has been prolonged to 900 seconds.
With a lower top, the recommended sponge layer limit had to be adjusted too, which is now at 100Pa instead of 50Pa.