Input file calendar adjustment
As is, if we use an output file from a previous run which was using a
360-day, thirty-day-months calendar, the calendar attribute within the
file will be `”360”`. Unfortunately, that calendar does not work for
input files, if the model calendar is set to `”thirty_day_months”`.
This fix makes sure that a `360` calendar is treated as
`thirty_day_months`.