Add files via upload
Modified files that in combination with ../model/spectral_dynamics.f90, allow one to run the model in a zonally-symmetric setup (i.e., only keeping wave-0) or to truncate the model to retain only a chosen wavenumber band. See ../model/spectral_dynamics.f90 for details on truncating the model to chosen wavenumbers.